requirements.txt
pandas
python-docx
requests
python-dotenv
streamlit
openpyxl
